fix : manualLotto, autoLotto 팩토리 메소드 수정
