+++ src/test/java/lotto/domain/AutoLottoTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.util.LottoFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,7 +10,7 @@ public class AutoLottoTest {
     @DisplayName("자동 로또 생성 확인")
     @Test
     void create_auto_lotto() {
-        AutoLotto autoLotto = LottoFactory.createAutoLotto(6);
+        AutoLotto autoLotto = AutoLotto.createAutoLotto(6);
 
         assertThat(autoLotto.getAutoLotto()).hasSize(6);
     }
