+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.util.LottoFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -19,17 +18,15 @@ public class WinningLottoTest {
         BonusNumber bonusNumber = new BonusNumber(7);
         WinningLotto winningLotto = new WinningLotto(new Lotto(winningNumber), bonusNumber);
 
-
         assertThat(winningLotto).isEqualTo(new WinningLotto(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6)), new BonusNumber(7)));
     }
 
     @DisplayName("로또 당첨결과 확인")
     @Test
     void testEntireLottoMatching() {
-
         Lottos lottos = new Lottos(
-                LottoFactory.createManualLotto(Arrays.asList("1, 2, 3, 20, 21, 40", "1, 2, 20, 25, 29, 45")),
-                LottoFactory.createAutoLotto(0)
+                ManualLotto.createManualLotto(Arrays.asList("1, 2, 3, 20, 21, 40", "1, 2, 20, 25, 29, 45")),
+                AutoLotto.createAutoLotto(0)
         );
 
         WinningLotto winningLotto = new WinningLotto(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6)), new BonusNumber(20));
